extract functions fom component
