+++ app/components/capability/constants.js
@@ -5,3 +5,51 @@ export const UPDATE_QUESTION_SET = 'UPDATE_QUESTION_SET'
 export const GET_ALL_QUESTIONS = 'GET_ALL_QUESTIONS'
 export const GET_ALL_ANSWERS = 'GET_ALL_ANSWERS'
 export const UPDATE_ANSWERS = 'UPDATE_ANSWERS'
+
+export const capabilityQuestions = [
+	{
+		name: 'class5',
+		class5: {
+			german: {
+				number: 6,
+				short: 'german'
+			},
+			mathematics: {
+				number: 4,
+				short: 'math'
+			},
+			english: {
+				number: 4,
+				short: 'eng'
+			},
+			science: {
+				number: 4,
+				short: 'sci'
+			},
+			music: {
+				number: 3,
+				short: 'music'
+			},
+			art: {
+				number: 4,
+				short: 'art'
+			},
+			work: {
+				number: 4,
+				short: 'work'
+			},
+			sport: {
+				number: 4,
+				short: 'sport'
+			},
+			history: {
+				number: 4,
+				short: 'history'
+			},
+			general: {
+				number: 14,
+				short: 'general'
+			}
+		}
+	}
+]
