+++ app/components/subject/helpers/formHelpers.jsx
@@ -53,3 +53,55 @@ export const resolveHiddenInput = (filteredData, id) => (
 		<Input type="hidden" name="subjectId" data-id={id} />
 	</div>
 )
+
+export const filterSubjects = (chosenClass, subjectData) => {
+	if (_.isNil(subjectData) || _.isNil(chosenClass)) {
+		return []
+	}
+
+	const chosenSubjects = _.chain(subjectData)
+		.filter(['classroomId', chosenClass._id])
+		.orderBy(
+			['abbreviation'],
+			[subJ => subJ.abbreviation.toLowerCase()],
+			['asc']
+		)
+		.value()
+
+	return chosenSubjects
+}
+
+export const generateSubjectList = (filteredData, action) =>
+	filteredData.map((data, idx) => (
+		<Button
+			key={idx}
+			data-id={data._id}
+			className={`list-group-item list-group-item-action ${css.list_btn}`}
+			onClick={action}
+		>
+			{data.abbreviation}
+
+			<span className={`badge badge-warning badge-pill ${css.badge_number}`}>
+				{data.tests.length}
+			</span>
+		</Button>
+	))
+
+export const generateInputs = (t, { name, abbreviation, isInvalid }) => {
+	const formLabels = { name, abbreviation }
+	return _.keys(formLabels).map((data, idx) => (
+		<div key={idx} className={css.form_div}>
+			<Label className={css.form_label} htmlFor={`${data}Sid`}>
+				{t(`room.${data}`)}:
+			</Label>
+
+			<Input
+				name={data}
+				className="form-control"
+				data-id={`${data}Sid`}
+				type="text"
+				invalid={isInvalid && _.isEmpty(formLabels[data])}
+			/>
+		</div>
+	))
+}
