--- app/constants/capabilityQuestions.js
@@ -1,47 +1,0 @@
-export const capabilityQuestions = [
-	{
-		name: 'class5',
-		class5: {
-			german: {
-				number: 6,
-				short: 'german'
-			},
-			mathematics: {
-				number: 4,
-				short: 'math'
-			},
-			english: {
-				number: 4,
-				short: 'eng'
-			},
-			science: {
-				number: 4,
-				short: 'sci'
-			},
-			music: {
-				number: 3,
-				short: 'music'
-			},
-			art: {
-				number: 4,
-				short: 'art'
-			},
-			work: {
-				number: 4,
-				short: 'work'
-			},
-			sport: {
-				number: 4,
-				short: 'sport'
-			},
-			history: {
-				number: 4,
-				short: 'history'
-			},
-			general: {
-				number: 14,
-				short: 'general'
-			}
-		}
-	}
-]
